Added Dancing Elves and Cleaned Up Files
We added a button at the bottom of our application that displays
a gif of dancing elves. We also formated it accordingly, and added
in a button that would remove the gif if wanted.

We also went in and cleaned up/ removed files so that our overall
application would be cleaner and more efficent.